apps/api/requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
supabase>=2.3.0,<3.0.0
httpx>=0.25.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
anthropic>=0.8.0
openai>=1.6.0
Pillow>=10.1.0
